+++ package.json
@@ -20,6 +20,7 @@
     "eva-icons": "^1.1.3",
     "expo": "~37.0.3",
     "expo-updates": "~0.2.0",
+    "mongoose": "^5.9.17",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "~0.61.5",
